Pipeline - automated version of the process
This notebook uses Nipype and the process from the 'Skullstrip_with_ANTs_and_FSL' notebook to semi-automate the skullstrip. It will apply the process to all T1_MPRAGE volumes in the designated folder. It requires a study-specific template and a manually made (in ITK-snap) mask to co-register the individual sheep and bring the mask into individual sheep space to skullstrip. It does this in a serial manner on all the T1 files in the folder.