A pipeline to skullstrip fMRI data
The fMRI data needs skullstripped before use in FSL's FEAT tool. Here a similar pipeline is used as before (ANTs registration and FSL ApplyMask) with the addition of taking one image from the time series (the 'min' image in the timeseries) and dilating the mask after it has been co-registered to individual sheep space.